pr_comments: rename evm_backends -> backends
